change location string to int from float

+++ app/android/ISMRR/app/src/main/java/ml/entc19fyp/ismrr/CustomMapView.java
@@ -157,7 +157,7 @@ public class CustomMapView extends androidx.appcompat.widget.AppCompatImageView 
 
         @NonNull
         public String toString() {
-            return this.x + "," + this.y;
+            return (int)this.x + "," + (int)this.y;
         }
 
     }
